2022, day 4
At first, I tried to iterate through all numbers of a given sequence
and find overlaps there. Of course, it didn't work. After banging my
head to the wall for about an hout, I just decided to do a normal
comparison, with some prototyping in a notebook to understand which
start/end property to use.

Never underestimate the power of pen and paper.