2022, day 11
First time this year I got stuck in a problem, and it will probably be
a more common occurrence. Part 1 wasn't particularly hard, but part 2
gave me slightly off values despite being almost the same thing. After
a bit of reflection, I assumed the numbers got too big at some point
and were giving me integer overflow errors. I tried using the BigInt
object to solve this, but aparently there's a limit to how big a
BigInt can be, at least on Firefox. Ultimately, I had to look for a
hint on how to solve this, and so I found out about the modular
arithmetic trick. Seems so obvious in retrospect, but I guess it's not
a very intuitive solution at first.