Optimize findZeros by filtering useless cells
Neighbors with the same value as the original cell are redundant for
pathfinding, so we skip them. Neighbors with a value of 2 + the
original cell or higher can't be visited. Thus, the only useful
neighbors are the ones whose value is that of the original cell +
1. Discarding the aforementioned cells while looking for starting
values to solve part 2 gives a speed increase of about 250% compared
to code that checks every single cell whose value is 0 (or "a", whatever.)