2022, day 15 (part 1)
Didn't feel like solving this at night, so I started working on this
after waking up, modeling the problem on pen and paper. As it turns
out, taxicab geometry is pretty fun and interesting.

My first solution added every single coincidence to the map, so it
took a while to complete (about three seconds.) Of course, instead of
adding each individual position to a map and then counting the total,
one can just use two values, one being the start of a range with all
those coincidences, and one being the end. Then, you just return the
difference between those values. It's a bit hard to explain, but easy
to understand.

Now, a naive solution for part 2 took all my memory and a lot of time
to compute, thus I have to think how to solve it (or perhaps look up a
solution online.) How do you even do that one? Ranges as well?