2023, day 1
I don't know if I'm rusty or this day 1 was particularly difficult. I
mean, part 1 was rather easy, but I got stuck on part 2 for about two
hours. I looked up a solution online which replaced every spelled-out
numeral with the first letter, the number, and the last
letter ("three" becomes "t3e", pretty clever,) and after writing a
file with right and wrong outputs and looking at the diff, I realized
I was looking for a single instance of a substring (so, "twothreetwo"
would only mark a "two" at position zero, but not the last one at
position 8). Maybe it's cheating, but I ultimately fixed the solution
that I had already built. Was this puzzle unusually tricky for the
first day because of ChatGPT prompters?