2023, day 2
The puzzle itself wasn't very hard, but I may have went a bit
overboard by writing types for this. Eh, it looks ok, I think. Still,
took me a bit of time to complete because I was being a bit too
careful checking my outputs. I was about to do it in javascript, but
the .split.map mess I was about to write was a bit too confusing; I
think types helped with this, by the way. Oh, I got part 1 wrong early
on, but that's because I thought I had to sum all sets of cubes (the
ones separated by semicolons) per game, rather than checking each set
individually, and that gave me a lot of false negatives. To be fair,
the example input gave me the same answer.

Oh, and I added a day selector for the binary using a map of
functions in the entry point. Special considerations for a compiled
language, I guess.