2023, day 04
This day was surprisingly easy, considering the difficulty of the
previous challenges. At first I was going to use a regular expression
to parse the card, but I had some problems with the submatch system,
so I decided to use do it the old-fashioned way with split and a
tokenizer (a very useful function, by the way). No need to call
math.Pow when the powers are all powers of 2, just shift the bits.

For part 2, I was getting worried when I skimmed the text and saw the
number of cards is supposed to grow... Hopefully, I thought about it
for a minute and realized it could be easily done with a
hashmap. Well, I got a bit confused about indices and how many times
to add the numbers, so I moved that part to its own function and also
grabbed my notebook to sketch the algorithm. Yeah, it turns out good
practices are good.