2023, day 07
I lost a lot of time wondering why I got the wrong answer for part 1,
getting differnt results with SortFunc and SortStableFunc. It turns
out I had a typo on the map[byte]int, writing something like '5', '6',
'6', '7', etc. Other than that, it wasn't as hard as I thought it
would be, considering today is an odd day. Maybe I should have
generated the map on the function itself, it would have prevented that error.