2024, day 01
As tradition demands, this year I'm using a language I haven't used
before in an instance of Advent of Code. This year's language is C, an
extremely painful and verbose choice that I thought would be fun to
use because this last year I've been using it pretty frequently.

Problem wasn't particularly hard, but coding it in C took me a while,
since the language itself is somewhat low-level-ish, and I was
debugging a furtive out-of-bounds bug that disappeared with no
apparent reason. I'm not an experienced C programmer, okay? That means
I was hesitant to implement a hashmap structure, especially for the
very first problem. While the algorithm I used runs in quadratic time,
the input file is small enough to be processed in a few milliseconds.