2024, day 4
The yearly tradition of switching to Python came early, mostly because
I don't know what I was thinking when I began using C for this. I
actually did the previous one in Javascript, but haven't commited it
yet. I thought I would have to do some tedious grid manipulation for
this problem, but it's mostly repeating a bunch of string comparisons
with some transformations (transposition and finding diagonals, as
well as checking each 3x3 square and converting it to a 9-character
string.)